refactor: trim MainIcon component name blank
